+++ packages/management-api/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.51.0](https://github.com/zuohuadong/supacloud/compare/management-api-v0.50.34...management-api-v0.51.0) (2026-08-10)
+
+
+### Features
+
+* **admin:** report deployed platform versions ([#794](https://github.com/zuohuadong/supacloud/issues/794)) ([1ed128c](https://github.com/zuohuadong/supacloud/commit/1ed128cb8100872a9fd4a93a373ac1df67d3baf4))
+
+## [0.50.34](https://github.com/zuohuadong/supacloud/compare/management-api-v0.50.33...management-api-v0.50.34) (2026-08-10)
+
+
+### Bug Fixes
+
+* **upgrade:** pin Sigstore root for offline verification ([#789](https://github.com/zuohuadong/supacloud/issues/789)) ([4bb67ca](https://github.com/zuohuadong/supacloud/commit/4bb67caf318710d825ee7dd950bbd8a3be71db27))
+
 ## [0.50.33](https://github.com/zuohuadong/supacloud/compare/management-api-v0.50.32...management-api-v0.50.33) (2026-08-10)